ENH: Refactor file io for positionfixes
